plus icon on brick editor side adds brick to tutorial editor. pen icon on tutorial editor side, adapt the changes of the brick in the brick editor to the currently selected brick on tutorial side.
