+++ frontend/Cargo.toml
@@ -13,8 +13,6 @@ usvg = "0.47.0"
 tiny-skia = { version = "0.12.0", features = ["png-format"] }
 serde_json = "1.0.149"
 base64 = "0.22.1"
-
-[target.'cfg(target_arch = "wasm32")'.dependencies]
 wasm-bindgen = "0.2"
 wasm-bindgen-futures = "0.4"
 web-sys = { version = "0.3", features = ["Window", "Document", "Element", "EventTarget", "File", "FileList", "FileReader", "HtmlElement", "HtmlAnchorElement", "HtmlCanvasElement", "CanvasRenderingContext2d", "HtmlImageElement", "HtmlInputElement", "HtmlTextAreaElement", "DragEvent", "DataTransfer", "Blob", "BlobPropertyBag", "Url", "ResizeObserver", "ResizeObserverEntry"] }
