+++ frontend/src/app/editors/brick.rs
@@ -7,6 +7,8 @@ use crate::components::editor_group::EditorGroup;
 #[cfg(target_arch = "wasm32")]
 use crate::interfaces::brick::BrickState;
 #[cfg(target_arch = "wasm32")]
+use crate::interfaces::tutorial::{TutorialAction, TutorialViewState};
+#[cfg(target_arch = "wasm32")]
 use yew::prelude::*;
 
 #[cfg(target_arch = "wasm32")]
@@ -14,11 +16,20 @@ use yew::prelude::*;
 pub struct BrickEditorProps {
     pub brick: BrickState,
     pub dispatcher: UseReducerDispatcher<BrickState>,
+    pub tutorial_dispatcher: UseReducerDispatcher<TutorialViewState>,
 }
 
 #[cfg(target_arch = "wasm32")]
 #[function_component(BrickEditor)]
 pub fn brick_editor(props: &BrickEditorProps) -> Html {
+    let on_add_to_tutorial = {
+        let dispatcher = props.tutorial_dispatcher.clone();
+        let brick = props.brick.clone();
+        Callback::from(move |_: MouseEvent| {
+            dispatcher.dispatch(TutorialAction::AddBrick(brick.clone()));
+        })
+    };
+
     html! {
         <EditorGroup title="Brick Editor">
             <BrickPreviewView
@@ -28,6 +39,7 @@ pub fn brick_editor(props: &BrickEditorProps) -> Html {
                 <BrickSettingsView
                 brick={props.brick.clone()}
                 dispatcher={props.dispatcher.clone()}
+                {on_add_to_tutorial}
             />
         </EditorGroup>
     }
